+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Require file now search for parent directory containing the composer.json
+
 ## [0.1.0] - 2020-02-28
 
 ### Added